+++ src/main/java/subway/controller/ManageController.java
@@ -1,5 +1,8 @@
-package subway.domain;
+package subway.controller;
 
+import subway.domain.Line;
+import subway.domain.LineRepository;
+import subway.domain.StationRepository;
 import subway.view.InputView;
 import subway.view.OutputView;
