+++ src/main/java/subway/view/OutputView.java
@@ -1,8 +1,8 @@
 package subway.view;
 
+import subway.controller.ManageController;
 import subway.domain.Line;
 import subway.domain.LineRepository;
-import subway.domain.ManageController;
 import subway.domain.StationRepository;
 import subway.domain.function.Functionable;
